fix merge consequence post rename, move loadtest to intended location
